Move UserGravity calls out of Sph.cpp
